products/serializers.py
from rest_framework.serializers import ModelSerializer

from products.models import Product, Brand


class ProductSerializer(ModelSerializer):
    class Meta:
        model = Product
        fields = '__all__'


class BrandSerializer(ModelSerializer):
    class Meta:
        model = Brand
        fields = '__all__'
