requirements.txt
Django==3.2.12
djangorestframework==3.13.1
django-filter==21.1
markdown==3.3.6
Pillow==9.0.1